Implementing support for Windows versions in install, list-installed, reveal-installed, run, uninstall and update cmds
This is a breaking change that changes how installations are tracked. After updating to this version you'll need to manually reinstall all products. Your save games are fine (nothing is removed with this version, theo just changed how installations are tracked completely). Here's how to check what products are installed:

- Before updating to this version run `list-installed` and `list-wine-installed` and save both lists
- If you've already updated - navigate to theo state folder (%UserData%/theo) then metadata/installed-details. Take note of all ids listed under each operating system folder

After getting the list of installed games - run `theo install {id}`, where {id} is one of the ids you've collected above.

This version changes install heuristics. If operating system is not specified - it'll attempt to install current OS version (if available) or Windows version on the current OS. Same applies to run, uninstall, reveal-installed - they will attempt to detect which OS version is installed and work on that.

With this change there's no need for wine-* commands as Windows versions are handled by the respective commands and even automatically in most cases where native version is not available.

If several OS versions are installed at the same time - you will need to specify which one to run/uninstall/reveal-installed.

Likewise - `update -all` will operate on all installed versions automatically